feat: Add workaround to have edit links
Hugo theme only suports poiting to markdown `md` files which is not the
case as these files are exported during the build process and not
available through source control.

For this reason, this require a custom footer injection based on the
existing one that can handle it gracefully. However, given that the
`org` root folder is not simmetric to the build folder, the existing
file hierarchy had to change.

In summary:
* Moved existing "main" content to `docs` folder.
* Injecting custom `footer` to have "edit page" link.
* Moved the `introduction.org` content to the proper file.